refactor(test): moved specs to a separate directory
backend/src/user/user.controller.spec.ts -> backend/src/user/test/user.controller.spec.ts
backend/src/user/user.service.spec.ts -> backend/src/user/test/user.service.spec.ts
