+++ backend/src/user/test/user.controller.spec.ts
@@ -1,14 +1,16 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserController } from '../user.controller';
+import { UserService } from '../user.service';
+import { UserRepositoryFake } from '../../mocks/user.repository.fake';
 
 describe('UserController', () => {
   let controller: UserController;
+  let repository: UserRepositoryFake;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [UserController],
-      providers: [UserService],
+      providers: [UserService, { provide: 'UserRepository', useValue: repository }],
     }).compile();
 
     controller = module.get<UserController>(UserController);
